Make a failed Poll invocation detectable by the user
Presently there is no way to determine if a Poll was successfull or not.
Setting the badge to 0 will allow for a caller to observe that something
has happened. It is assumed that no 'real' client would have a zero badge
and so the absence of a badge indicates failure